+++ R/esp-get-grid-ESDAC.R
@@ -29,7 +29,7 @@
 #' @examplesIf esp_check_access()
 #' \dontrun{
 #' grid <- esp_get_grid_ESDAC()
-#' esp <- esp_get_country(moveCAN = FALSE)
+#' esp <- esp_get_spain(moveCAN = FALSE)
 #'
 #' library(ggplot2)
 #'