+++ tests/testthat/test_local/test-wmts-providers-png.R
@@ -28,15 +28,21 @@ test_that("Test WMTS png", {
 
   cala <- esp_get_capimun(munic = "Calahorra", epsg = 3857, region = "La Rioja")
   cala <- cala |> sf::st_buffer(dist = 1000)
-
+  fails <- NULL
   for (n in all_n) {
     tile <- try(
       esp_get_tiles(cala, type = n, cache_dir = cdir),
       silent = TRUE
     )
     if (!inherits(tile, "try-error")) {
+      expect_true(!is.null(ensure_null(terra::crs(tile))))
+
       expect_snapshot_file(save_png(tile), paste0(n, ".png"))
+    } else {
+      fails <- c(fails, n)
     }
   }
+  expect_snapshot(fails)
+
   unlink(cdir, force = TRUE, recursive = TRUE)
 })